refactor exec flag
