+++ cmd/nerdctl/container_exec.go
@@ -17,25 +17,12 @@
 package main
 
 import (
-	"context"
 	"errors"
-	"fmt"
-	"io"
-	"os"
 
-	"github.com/containerd/console"
 	"github.com/containerd/containerd"
-	"github.com/containerd/containerd/cio"
-	"github.com/containerd/containerd/cmd/ctr/commands"
-	"github.com/containerd/containerd/cmd/ctr/commands/tasks"
+	"github.com/containerd/nerdctl/pkg/api/types"
 	"github.com/containerd/nerdctl/pkg/clientutil"
-	"github.com/containerd/nerdctl/pkg/flagutil"
-	"github.com/containerd/nerdctl/pkg/idgen"
-	"github.com/containerd/nerdctl/pkg/idutil/containerwalker"
-	"github.com/containerd/nerdctl/pkg/taskutil"
-	"github.com/opencontainers/runtime-spec/specs-go"
-
-	"github.com/sirupsen/logrus"
+	"github.com/containerd/nerdctl/pkg/cmd/container"
 	"github.com/spf13/cobra"
 )
 
@@ -64,8 +51,74 @@ func newExecCommand() *cobra.Command {
 	return execCommand
 }
 
-func execAction(cmd *cobra.Command, args []string) error {
+func processExecCommandOptions(cmd *cobra.Command) (types.ContainerExecOptions, error) {
 	globalOptions, err := processRootCmdFlags(cmd)
+	if err != nil {
+		return types.ContainerExecOptions{}, err
+	}
+
+	flagI, err := cmd.Flags().GetBool("interactive")
+	if err != nil {
+		return types.ContainerExecOptions{}, err
+	}
+	flagT, err := cmd.Flags().GetBool("tty")
+	if err != nil {
+		return types.ContainerExecOptions{}, err
+	}
+	flagD, err := cmd.Flags().GetBool("detach")
+	if err != nil {
+		return types.ContainerExecOptions{}, err
+	}
+
+	if flagI {
+		if flagD {
+			return types.ContainerExecOptions{}, errors.New("currently flag -i and -d cannot be specified together (FIXME)")
+		}
+	}
+
+	if flagT {
+		if flagD {
+			return types.ContainerExecOptions{}, errors.New("currently flag -t and -d cannot be specified together (FIXME)")
+		}
+	}
+
+	workdir, err := cmd.Flags().GetString("workdir")
+	if err != nil {
+		return types.ContainerExecOptions{}, err
+	}
+
+	envFile, err := cmd.Flags().GetStringSlice("env-file")
+	if err != nil {
+		return types.ContainerExecOptions{}, err
+	}
+	env, err := cmd.Flags().GetStringArray("env")
+	if err != nil {
+		return types.ContainerExecOptions{}, err
+	}
+	privileged, err := cmd.Flags().GetBool("privileged")
+	if err != nil {
+		return types.ContainerExecOptions{}, err
+	}
+	user, err := cmd.Flags().GetString("user")
+	if err != nil {
+		return types.ContainerExecOptions{}, err
+	}
+
+	return types.ContainerExecOptions{
+		GOptions:    globalOptions,
+		TTY:         flagT,
+		Interactive: flagI,
+		Detach:      flagD,
+		Workdir:     workdir,
+		Env:         env,
+		EnvFile:     envFile,
+		Privileged:  privileged,
+		User:        user,
+	}, nil
+}
+
+func execAction(cmd *cobra.Command, args []string) error {
+	options, err := processExecCommandOptions(cmd)
 	if err != nil {
 		return err
 	}
@@ -76,199 +129,14 @@ func execAction(cmd *cobra.Command, args []string) error {
 		newArg = append(newArg, args[2:]...)
 		args = newArg
 	}
-	client, ctx, cancel, err := clientutil.NewClient(cmd.Context(), globalOptions.Namespace, globalOptions.Address)
+
+	client, ctx, cancel, err := clientutil.NewClient(cmd.Context(), options.GOptions.Namespace, options.GOptions.Address)
 	if err != nil {
 		return err
 	}
 	defer cancel()
 
-	walker := &containerwalker.ContainerWalker{
-		Client: client,
-		OnFound: func(ctx context.Context, found containerwalker.Found) error {
-			if found.MatchCount > 1 {
-				return fmt.Errorf("multiple IDs found with provided prefix: %s", found.Req)
-			}
-			return execActionWithContainer(ctx, cmd, args, found.Container, client)
-		},
-	}
-	req := args[0]
-	n, err := walker.Walk(ctx, req)
-	if err != nil {
-		return err
-	} else if n == 0 {
-		return fmt.Errorf("no such container %s", req)
-	}
-	return nil
-}
-
-func execActionWithContainer(ctx context.Context, cmd *cobra.Command, args []string, container containerd.Container, client *containerd.Client) error {
-	flagI, err := cmd.Flags().GetBool("interactive")
-	if err != nil {
-		return err
-	}
-	flagT, err := cmd.Flags().GetBool("tty")
-	if err != nil {
-		return err
-	}
-	flagD, err := cmd.Flags().GetBool("detach")
-	if err != nil {
-		return err
-	}
-
-	if flagI {
-		if flagD {
-			return errors.New("currently flag -i and -d cannot be specified together (FIXME)")
-		}
-	}
-
-	if flagT {
-		if flagD {
-			return errors.New("currently flag -t and -d cannot be specified together (FIXME)")
-		}
-	}
-
-	pspec, err := generateExecProcessSpec(ctx, cmd, args, container, client)
-	if err != nil {
-		return err
-	}
-
-	task, err := container.Task(ctx, nil)
-	if err != nil {
-		return err
-	}
-	var (
-		ioCreator cio.Creator
-		in        io.Reader
-		stdinC    = &taskutil.StdinCloser{
-			Stdin: os.Stdin,
-		}
-	)
-
-	if flagI {
-		in = stdinC
-	}
-	cioOpts := []cio.Opt{cio.WithStreams(in, os.Stdout, os.Stderr)}
-	if flagT {
-		cioOpts = append(cioOpts, cio.WithTerminal)
-	}
-	ioCreator = cio.NewCreator(cioOpts...)
-
-	execID := "exec-" + idgen.GenerateID()
-	process, err := task.Exec(ctx, execID, pspec, ioCreator)
-	if err != nil {
-		return err
-	}
-	stdinC.Closer = func() {
-		process.CloseIO(ctx, containerd.WithStdinCloser)
-	}
-	// if detach, we should not call this defer
-	if !flagD {
-		defer process.Delete(ctx)
-	}
-
-	statusC, err := process.Wait(ctx)
-	if err != nil {
-		return err
-	}
-
-	var con console.Console
-	if flagT {
-		con = console.Current()
-		defer con.Reset()
-		if err := con.SetRaw(); err != nil {
-			return err
-		}
-	}
-	if !flagD {
-		if flagT {
-			if err := tasks.HandleConsoleResize(ctx, process, con); err != nil {
-				logrus.WithError(err).Error("console resize")
-			}
-		} else {
-			sigc := commands.ForwardAllSignals(ctx, process)
-			defer commands.StopCatch(sigc)
-		}
-	}
-
-	if err := process.Start(ctx); err != nil {
-		return err
-	}
-	if flagD {
-		return nil
-	}
-	status := <-statusC
-	code, _, err := status.Result()
-	if err != nil {
-		return err
-	}
-	if code != 0 {
-		return fmt.Errorf("exec failed with exit code %d", code)
-	}
-	return nil
-}
-
-func generateExecProcessSpec(ctx context.Context, cmd *cobra.Command, args []string, container containerd.Container, client *containerd.Client) (*specs.Process, error) {
-	spec, err := container.Spec(ctx)
-	if err != nil {
-		return nil, err
-	}
-	userOpts, err := generateUserOpts(cmd)
-	if err != nil {
-		return nil, err
-	}
-	if userOpts != nil {
-		c, err := container.Info(ctx)
-		if err != nil {
-			return nil, err
-		}
-		for _, opt := range userOpts {
-			if err := opt(ctx, client, &c, spec); err != nil {
-				return nil, err
-			}
-		}
-	}
-
-	pspec := spec.Process
-	flagT, err := cmd.Flags().GetBool("tty")
-	if err != nil {
-		return nil, err
-	}
-	pspec.Terminal = flagT
-	pspec.Args = args[1:]
-
-	workdir, err := cmd.Flags().GetString("workdir")
-	if err != nil {
-		return nil, err
-	}
-	if workdir != "" {
-		pspec.Cwd = workdir
-	}
-	envFile, err := cmd.Flags().GetStringSlice("env-file")
-	if err != nil {
-		return nil, err
-	}
-	env, err := cmd.Flags().GetStringArray("env")
-	if err != nil {
-		return nil, err
-	}
-	envs, err := generateEnvs(envFile, env)
-	if err != nil {
-		return nil, err
-	}
-	pspec.Env = flagutil.ReplaceOrAppendEnvValues(pspec.Env, envs)
-
-	privileged, err := cmd.Flags().GetBool("privileged")
-	if err != nil {
-		return nil, err
-	}
-	if privileged {
-		err = setExecCapabilities(pspec)
-		if err != nil {
-			return nil, err
-		}
-	}
-
-	return pspec, nil
+	return container.Exec(ctx, client, args, options)
 }
 
 func execShellComplete(cmd *cobra.Command, args []string, toComplete string) ([]string, cobra.ShellCompDirective) {
